+++ src/main/java/com/demo/community/controller/PublishController.java
@@ -31,7 +31,7 @@ public class PublishController {
             @RequestParam(name="title",required = false)String title,
             @RequestParam(name="description",required = false) String description,
             @RequestParam(name="tag",required = false)String tag,
-            @RequestParam(name="id",required = false)Integer id,
+            @RequestParam(name="id",required = false)Long id,
             HttpServletRequest request,
             Model model
     ){
@@ -68,7 +68,7 @@ public class PublishController {
 
     @GetMapping("/publish/{id}")
     public String edit(
-            @PathVariable(name = "id") Integer id,
+            @PathVariable(name = "id") Long id,
             Model model,
             HttpServletRequest request
     ){
