+++ src/main/resources/db/migration/V3__Create_question_table.sql
@@ -1,11 +1,11 @@
 create table question (
 
-    id int auto_increment primary key not null,
+    id bigint auto_increment primary key not null,
     title varchar(50),
     description text,
     gmt_create bigint,
     gmt_modified bigint,
-    creator_id int,
+    creator_id bigint,
     comment_count int default 0,
     view_count int default 0,
     like_count int default 0,
